Refactor realize_reduction, move to separate file
